requirements.txt
torch==1.6.0
transformers==3.0.2
seqeval==0.0.12
pytorch-crf==0.7.2
Flask==1.0.3
gunicorn==19.9.0
